+++ src/contentsharecontroller/DefaultContentShareController.ts
@@ -1,6 +1,7 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import AudioProfile from '../audioprofile/AudioProfile';
 import AudioVideoController from '../audiovideocontroller/AudioVideoController';
 import AudioVideoObserver from '../audiovideoobserver/AudioVideoObserver';
 import ContentShareObserver from '../contentshareobserver/ContentShareObserver';
@@ -43,6 +44,10 @@ export default class DefaultContentShareController
   ) {
     this.contentAudioVideo.addObserver(this);
     this.setupContentShareEvents();
+  }
+
+  setContentAudioProfile(audioProfile: AudioProfile): void {
+    this.contentAudioVideo.setAudioProfile(audioProfile);
   }
 
   async startContentShare(stream: MediaStream): Promise<void> {