+++ test/sdp/DefaultSDP.test.ts
@@ -87,6 +87,64 @@ describe('DefaultSDP', () => {
     });
   });
 
+  describe('withAudioMaxAverageBitrate', () => {
+    it('adds the bitrate parameter', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(64000);
+      expect(sdpB.sdp).to.equal(
+        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITH_AUDIO_BITRATE.replace('$VALUE', '64000')
+      );
+    });
+
+    it('updates an old bitrate parameter', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(64000).withAudioMaxAverageBitrate(48000);
+      expect(sdpB.sdp).to.equal(
+        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITH_AUDIO_BITRATE.replace('$VALUE', '48000')
+      );
+    });
+
+    it('returns same SDP if null passed in', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(null);
+      expect(sdpB.sdp).to.equal(sdpA.sdp);
+    });
+
+    it('returns same SDP if zero passed in', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(0);
+      expect(sdpB.sdp).to.equal(sdpA.sdp);
+    });
+
+    it('returns same SDP if undefined passed in', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(undefined);
+      expect(sdpB.sdp).to.equal(sdpA.sdp);
+    });
+
+    it('clamps to the lowest bitrate', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(1);
+      expect(sdpB.sdp).to.equal(
+        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITH_AUDIO_BITRATE.replace(
+          '$VALUE',
+          `${DefaultSDP.rfc7587LowestBitrate}`
+        )
+      );
+    });
+
+    it('clamps to the highest bitrate', () => {
+      const sdpA = new DefaultSDP(SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO);
+      const sdpB = sdpA.withAudioMaxAverageBitrate(Infinity);
+      expect(sdpB.sdp).to.equal(
+        SDPMock.LOCAL_OFFER_WITH_AUDIO_VIDEO_WITH_AUDIO_BITRATE.replace(
+          '$VALUE',
+          `${DefaultSDP.rfc7587HighestBitrate}`
+        )
+      );
+    });
+  });
+
   describe('withBandwidthRestriction', () => {
     it('adds bandwidth line for Chrome', () => {
       const sdpA = new DefaultSDP(SDPMock.VIDEO_HOST_AUDIO_VIDEO_ANSWER);